Create Knock-knock joke communication
This is a code to program a TCP client that  connects to a server and allows a user to tell the server knock-knock jokes. Communication with the server will be according to the attached Knock Knock Protocol, an application layer protocol that connects to a TCP socket running on port 8989. 
The hostname server you will connect to is ec2-34-218-41-170.us-west-2.compute.amazonaws.com.